+++ src/components/Products/AddingProductPageOne/AddingProduct/AddingProduct.js
@@ -21,6 +21,11 @@ import TypesDropUp from "../MobileDropUpSides/TypesDropUp/TypesDropUp";
 import AllSizeListForWear from "../../../../hook/AllSizeListForWear/AllSizeListForWear";
 import { FaRandom } from "react-icons/fa";
 import { useQuery } from "@tanstack/react-query";
+import HeadWearAdd from "./Details/HeadWear/HeadWearAdd";
+import OutWearAdd from "./Details/OutWear/OutWearAdd";
+import UnderWearAdd from "../../AddingProductPageOne copy/AddingProduct/Details/UnderWear copy/underWearAdd";
+import AccessoriesAdd from "./Details/Accessories/AccessoriesAdd";
+import ShoesAdd from "./Details/Shoes/ShoesAdd";
 
 const AddingProduct = () => {
 
@@ -134,931 +139,6 @@ const AddingProduct = () => {
     console.log(`switch to ${checked}`);
   };
 
-
-  // --------------------------------------Возрастная категория Артикул Обхват Талии Цена Возраст
-
-
-  // Hats
-  const contentHat = (
-    <div className="w-[520px] h-fit">
-      <div
-        className={`w-full h-fit flex flex-col items-center justify-center not-italic cursor-pointer font-AeonikProMedium text-sm leading-4 text-center hover:bg-bgColor`}
-      >
-        <div className="w-full flex justify-center  gap-x-10 px-3 pt-5">
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Обхват головы
-              <span className="text-sm text-textLightColor ml-[6px]">(см)</span>
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-
-            <div className="w-full flex items-center mt-[7px]">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[55px] h-[38px] text-center border border-borderColor px-2 rounded-lg   outline-none font-AeonikProRegular "
-                  placeholder="Мин"
-                />
-              </div>
-              <span className="mx-[5px]"><LineIcon /></span>
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[55px] h-[38px] text-center border border-borderColor px-2 rounded-lg  font-AeonikProRegular  outline-none"
-                  placeholder="Макс"
-                />
-              </div>
-            </div>
-          </div>
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center justify-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              One Size
-              <span className="text-sm text-textLightColor ml-[6px]">(см)</span>
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="flex items-center justify-center mt-[10px]">
-              <Switch className="bg-[#8B8B8B]" onChange={onChangeSwitch} />
-            </div>
-          </div>
-          <div className="w-fit flex flex-col items-center">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Количество
-              <span className="text-sm text-textLightColor ml-[6px]">(см)</span>
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="flex items-start justify-between ">
-              <input
-                type="text"
-                className="inputStyle w-[60px] h-[38px] text-center border border-borderColor px-5  rounded-lg  font-AeonikProRegular "
-              />
-            </div>
-          </div>
-        </div>
-        <div className="w-full flex flex-row px-3 gap-x-[11px] md:gap-x-[30px] mb-[15px] md:mt-[15px]">
-          <div className="w-1/2 flex items-center gap-x-[25px]">
-            <div className="w-fit hidden md:flex flex-col items-start">
-              <div className="flex items-center justify-center  mb-2 ll:mb-[10px]">
-                <label
-                  htmlFor=""
-                  className="flex items-center text-[14px] ll:text-base text-mobileTextColor  ll:font-AeonikProMedium font-AeonikProRegular">
-                  Возраст
-                </label>
-              </div>
-              <div className="w-full flex items-center">
-                <input
-                  type="text"
-                  className="inputStyle w-[58px] h-[42px] text-center fon border border-borderColor rounded-lg px-[12px]  outline-none "
-                  placeholder=""
-                />
-              </div>
-            </div>
-            <div className="w-full md:w-[90%]">
-              <div className="flex items-center mb-2 ll:mb-[10px] ">
-                <label
-                  htmlFor=""
-                  className="flex items-center text-[14px] ll:text-base text-mobileTextColor ll:font-AeonikProMedium font-AeonikProRegular">
-                  Цена
-                </label>
-                <span className="ml-[5px]">
-                  <StarLabel />
-                </span>
-              </div>
-              <div className="w-full h-[40px] flex items-center border border-borderColor px-3 py-[6px] rounded-lg text-xs">
-                <input
-                  type="text"
-                  placeholder="0"
-                  className="inputStyle w-[70%] font-AeonikProMedium outline-none "
-                />
-                <span className="text-textLightColor ml-[10px] text-xs md:text-base font-AeonikProRegular">
-                  сум
-                </span>
-              </div>
-            </div>
-          </div>
-          <div className="w-1/2 flex flex-col items-start">
-            <div className="flex items-center justify-center mb-2 ll:mb-[10px] ">
-              <label
-                htmlFor=""
-                className="flex items-center text-[14px] ll:text-base text-mobileTextColor  ll:font-AeonikProMedium font-AeonikProRegular">
-
-                Скидка
-              </label>
-              <span className="font-AeonikProMedium text-[10px] md:text-[13px] text-textLightColor ml-[5px]">
-                (необязательно)
-              </span>
-            </div>
-            <div className="w-full flex items-center justify-center">
-              <div className="w-full flex items-center gap-x-1">
-                <div className="w-[40%] md:w-[72px] flex items-start">
-                  <div className="w-full h-10 flex items-center justify-center border border-borderColor rounded-lg px-[10px] md:px-3 py-[8px]">
-                    <input
-                      type="text"
-                      placeholder="0"
-                      className="inputStyle w-[70%] font-AeonikProMedium text-start outline-none "
-                    />
-                    <span className="text-textLightColor ml-2">%</span>
-                  </div>
-                </div>
-                <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-                <div className="w-[60%] md:w-[75%] flex items-center">
-                  <div className="w-full h-[40px] flex items-center justify-between border border-borderColor px-3 py-[6px] rounded-lg text-xs">
-                    <input
-                      type="text"
-                      placeholder="0"
-                      className="inputStyle w-[75%] font-AeonikProMedium outline-none "
-                    />
-                    <span className="text-textLightColor ml-[10px] text-xs md:text-base font-AeonikProRegular">
-                      сум
-                    </span>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <button className="w-full flex items-end justify-end text-lg text-textBlueColor font-AeonikProMedium pr-1">
-          готово
-        </button>
-      </div>
-    </div>
-  );
-  // Outerwear bor 
-  const contentOutwear = (
-    <div className="w-[855px] h-fit">
-      <div
-        className={`w-full h-fit flex flex-col items-center not-italic cursor-pointer font-AeonikProMedium text-sm leading-4 text-center hover:bg-bgColor`}
-      >
-        <div className="w-full flex  gap-x-10 px-3 pt-5">
-          <div className="w-[20%] flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Обхват Груди
-              <span className="text-sm text-textLightColor ml-[6px]">(см)</span>
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="flex items-center">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] text-center h-[38px] border border-borderColor px-3  rounded-lg  font-AeonikProRegular "
-                  placeholder="Мин"
-                />
-              </div>
-              <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] text-center h-[38px] border border-borderColor px-3  rounded-lg font-AeonikProRegular "
-                  placeholder="Макс"
-                />
-              </div>
-            </div>
-          </div>
-          <div className="w-[20%] flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Размер
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="flex items-center">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] text-center h-[38px] border border-borderColor px-3  rounded-lg font-AeonikProRegular "
-                  placeholder="Мин"
-                />
-              </div>
-              <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] text-center h-[38px] border border-borderColor px-3  rounded-lg font-AeonikProRegular "
-                  placeholder="Макс"
-                />
-              </div>
-            </div>
-          </div>
-          <div className="w-[53%] flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-              Обхват Размер
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <AllSizeListForWear />
-          </div>
-        </div>
-        <div className="w-full flex justify-start items-center gap-x-10 px-3 pt-5">
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Обхват Талии
-              <span className="text-sm text-textLightColor ml-[6px]">(см)</span>
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="flex items-center">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] h-[38px]  text-center border border-borderColor px-2 md:px-3  rounded-lg   font-AeonikProRegular "
-                  placeholder="Мин"
-                />
-              </div>
-              <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] h-[38px]  text-center border border-borderColor px-2 md:px-3  rounded-lg  font-AeonikProRegular "
-                  placeholder="Макс"
-                />
-              </div>
-            </div>
-          </div>
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Обхват Бедер
-            </p>
-            <div className="flex items-center">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] h-[38px]  text-center border border-borderColor px-2 md:px-3  rounded-lg   font-AeonikProRegular "
-                  placeholder="Мин"
-                />
-              </div>
-              <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] h-[38px] text-center border border-borderColor px-2 md:px-3  rounded-lg  font-AeonikProRegular "
-                  placeholder="Макс"
-                />
-              </div>
-            </div>
-          </div>
-          <div className="w-fit flex flex-col md:ml-5">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Количество
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="flex items-start justify-between ">
-              <input
-                type="text"
-                className="inputStyle  w-[60px] h-[38px] text-center border border-borderColor px-5  rounded-lg  font-AeonikProRegular "
-              />
-            </div>
-          </div>
-        </div>
-        <div className="w-full flex flex-row px-3 gap-x-[11px] md:pt-5 md:gap-x-[20px] mb-[15px]">
-          <div className="w-fit flex items-center gap-x-[25px]">
-            <div className="w-fit hidden md:flex flex-col items-start">
-              <div className="flex items-center justify-center ">
-                <label
-                  htmlFor=""
-                  className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-                  Возраст
-                </label>
-                {/* <span className="ml-[5px]">
-                  <StarLabel />
-                </span> */}
-              </div>
-              <div className="w-fit flex items-center">
-                <input
-                  type="text"
-                  className="inputStyle w-[58px] h-[42px] text-center fon border border-borderColor rounded-lg   outline-none"
-                  placeholder=""
-                />
-              </div>
-            </div>
-            <div className="w-full md:w-[55%]">
-              <div className="flex items-center mb-2 ll:mb-[10px] ">
-                <label
-                  htmlFor=""
-                  className="flex items-center text-[14px] ll:text-base text-mobileTextColor  ll:font-AeonikProMedium font-AeonikProRegular">
-                  Цена
-                </label>
-                <span className="ml-[5px]">
-                  <StarLabel />
-                </span>
-              </div>
-              <div className="w-full h-[40px] flex items-center border border-borderColor px-3 py-[6px] rounded-lg text-xs">
-                <input
-                  type="text"
-                  placeholder="0"
-                  className="inputStyle w-[70%] font-AeonikProMedium outline-none "
-                />
-                <span className="text-textLightColor ml-[10px] text-xs md:text-base font-AeonikProRegular">
-                  сум
-                </span>
-              </div>
-            </div>
-          </div>
-          <div className="w-fit flex flex-col items-start">
-            <div className="flex items-center justify-center mb-2 ll:mb-[10px] ">
-              <label
-                htmlFor=""
-                className="flex items-center text-[14px] ll:text-base text-mobileTextColor ll:font-AeonikProMedium font-AeonikProRegular">
-
-                Скидка
-              </label>
-              <span className="font-AeonikProMedium text-[10px] md:text-[13px] text-textLightColor ml-[5px]">
-                (необязательно)
-              </span>
-            </div>
-            <div className="w-full flex items-center justify-center">
-              <div className="w-full flex items-center gap-x-1">
-                <div className="w-[40%] md:w-[72px] flex items-start">
-                  <div className="w-full h-10 flex items-center justify-center border border-borderColor rounded-lg px-[10px] md:px-3 py-[8px]">
-                    <input
-                      type="text"
-                      placeholder="0"
-                      className="inputStyle w-[70%] font-AeonikProMedium text-start outline-none "
-                    />
-                    <span className="text-textLightColor ml-2">%</span>
-                  </div>
-                </div>
-                <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-                <div className="w-[60%] md:w-[75%] flex items-center">
-                  <div className="w-full h-[40px] flex items-center justify-between border border-borderColor px-3 py-[6px] rounded-lg text-xs">
-                    <input
-                      type="text"
-                      placeholder="0"
-                      className="inputStyle w-[75%] font-AeonikProMedium outline-none "
-                    />
-                    <span className="text-textLightColor ml-[10px] text-xs md:text-base font-AeonikProRegular">
-                      сум
-                    </span>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <button className="w-full flex items-end justify-end text-lg text-textBlueColor font-AeonikProMedium pr-1">
-          готово
-        </button>
-      </div>
-    </div>
-  );
-  // Underwear bor Цена
-  const contentUnderWear = (
-    <div className="w-[840px] h-fit">
-      <div
-        className={`w-full h-fit flex flex-col items-center justify-center not-italic cursor-pointer font-AeonikProMedium text-sm leading-4 text-center hover:bg-bgColor`}
-      >
-        <div className="w-full flex gap-x-10 px-3 pt-5">
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Обхват Талии
-              <span className="text-sm text-textLightColor ml-[6px]">(см)</span>
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="flex items-center">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] h-[38px] text-center border border-borderColor px-3  rounded-lg   font-AeonikProRegular "
-                  placeholder="Мин"
-                />
-              </div>
-              <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] h-[38px] text-center border border-borderColor px-3  rounded-lg  font-AeonikProRegular "
-                  placeholder="Макс"
-                />
-              </div>
-            </div>
-          </div>
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Размер
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="flex items-center justify-between gap-x-1">
-              <div className="flex items-center">
-                <div className="flex flex-col">
-                  <input
-                    type="text"
-                    className="inputStyle w-[60px] text-center h-[38px] border border-borderColor px-3  rounded-lg   font-AeonikProRegular "
-                    placeholder="Мин"
-                  />
-                </div>
-                <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-                <div className="flex flex-col">
-                  <input
-                    type="text"
-                    className="inputStyle w-[60px] text-center h-[38px] border border-borderColor px-3  rounded-lg  font-AeonikProRegular "
-                    placeholder="Макс"
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="w-[53%] flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-              Обхват Размер
-            </p>
-            <AllSizeListForWear />
-          </div>
-        </div>
-        <div className="w-full flex gap-x-10 px-3 pt-5">
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Размер Бедер
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="flex items-center">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] h-[38px] text-center border border-borderColor px-3  rounded-lg   font-AeonikProRegular "
-                  placeholder="Мин"
-                />
-              </div>
-              <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] h-[38px] text-center border border-borderColor px-3  rounded-lg  font-AeonikProRegular "
-                  placeholder="Макс"
-                />
-              </div>
-            </div>
-          </div>
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Рост
-            </p>
-            <div className="flex items-center justify-between gap-x-1">
-              <div className="flex items-center">
-                <div className="flex flex-col">
-                  <input
-                    type="text"
-                    className="inputStyle w-[60px] text-center h-[38px] border border-borderColor px-3  rounded-lg   font-AeonikProRegular "
-                    placeholder="Мин"
-                  />
-                </div>
-                <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-                <div className="flex flex-col">
-                  <input
-                    type="text"
-                    className="inputStyle w-[60px] text-center h-[38px] border border-borderColor px-3  rounded-lg  font-AeonikProRegular "
-                    placeholder="Макс"
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="w-fit flex flex-col md:ml-[14px]">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Количество
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="flex items-start justify-between ">
-              <input
-                type="text"
-                className="inputStyle w-[60px] h-[38px] text-center border border-borderColor px-5  rounded-lg  font-AeonikProRegular "
-              />
-            </div>
-          </div>
-        </div>
-        <div className="w-full flex flex-row px-3 pt-5 gap-x-[11px] md:gap-x-[20px] mb-[15px]">
-          <div className="w-fit flex items-center gap-x-[25px]">
-            <div className="w-fit hidden md:flex flex-col items-start">
-              <div className="flex items-center justify-center  mb-2 ll:mb-[10px]">
-                <label
-                  htmlFor=""
-                  className="flex items-center text-[14px] ll:text-base text-mobileTextColor  ll:font-AeonikProMedium font-AeonikProRegular">
-
-                  Возраст
-                </label>
-                {/* <span className="ml-[5px]">
-                  <StarLabel />
-                </span> */}
-              </div>
-              <div className="w-fit flex items-center">
-                <input
-                  type="text"
-                  className=" inputStyle w-[58px] h-[42px] text-center fon border border-borderColor rounded-lg px-[12px]  outline-none"
-                  placeholder=""
-                />
-              </div>
-            </div>
-            <div className="w-full md:w-[55%]">
-              <div className="flex items-center  mb-2 ll:mb-[10px]">
-                <label
-                  htmlFor=""
-                  className="flex items-center text-[14px] ll:text-base text-mobileTextColor  ll:font-AeonikProMedium font-AeonikProRegular">
-
-                  Цена
-                </label>
-                <span className="ml-[5px]">
-                  <StarLabel />
-                </span>
-              </div>
-              <div className="w-full h-[40px] flex items-center border border-borderColor px-3 py-[6px] rounded-lg text-xs">
-                <input
-                  type="text"
-                  placeholder="0"
-                  className="inputStyle w-[70%] font-AeonikProMedium outline-none"
-                />
-                <span className="text-textLightColor ml-[10px] text-xs md:text-base font-AeonikProRegular">
-                  сум
-                </span>
-              </div>
-            </div>
-          </div>
-          <div className="w-fit flex flex-col items-start">
-            <div className="flex items-center justify-center mb-2 ll:mb-[10px] ">
-              <label
-                htmlFor=""
-                className="flex items-center text-[14px] ll:text-base text-mobileTextColor ll:font-AeonikProMedium font-AeonikProRegular">
-
-                Скидка
-              </label>
-              <span className="font-AeonikProMedium text-[10px] md:text-[13px] text-textLightColor ml-[5px]">
-                (необязательно)
-              </span>
-            </div>
-            <div className="w-full flex items-center justify-center">
-              <div className="w-full flex items-center gap-x-1">
-                <div className="w-[40%] md:w-[72px] flex items-start">
-                  <div className="w-full h-10 flex items-center justify-center border border-borderColor rounded-lg px-[10px] md:px-3 py-[8px]">
-                    <input
-                      type="text"
-                      placeholder="0"
-                      className="inputStyle w-[70%] font-AeonikProMedium text-start outline-none "
-                    />
-                    <span className="text-textLightColor ml-2">%</span>
-                  </div>
-                </div>
-                <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-                <div className="w-[60%] md:w-[75%] flex items-center">
-                  <div className="w-full h-[40px] flex items-center justify-between border border-borderColor px-3 py-[6px] rounded-lg text-xs">
-                    <input
-                      type="text"
-                      placeholder="0"
-                      className="inputStyle w-[75%] font-AeonikProMedium outline-none "
-                    />
-                    <span className="text-textLightColor ml-[10px] text-xs md:text-base font-AeonikProRegular">
-                      сум
-                    </span>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <button className="w-full flex items-end justify-end text-lg text-textBlueColor font-AeonikProMedium pr-1">
-          готово
-        </button>
-      </div>
-    </div>
-  );
-  // Shoes
-  const contentShoes = (
-    <div className="w-fit h-fit">
-      <div
-        className={`w-full h-fit flex flex-col items-center justify-center not-italic cursor-pointer font-AeonikProMedium text-sm leading-4 text-center hover:bg-bgColor`}
-      >
-        <div className="w-full flex gap-x-10 px-3 pt-5">
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Размер
-              <span className="ml-[5px]">
-                <StarLabel />
-              </span>
-            </p>
-            <div className="w-[65px] flex items-center justify-between gap-x-1">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-full text-start h-[40px] border border-borderColor px-3  rounded-lg   font-AeonikProRegular "
-                />
-              </div>
-            </div>
-          </div>
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Длина Стопы
-              <span className="text-sm text-textLightColor ml-[6px]">(см)</span>
-            </p>
-            <div className="flex items-center gap-x-1">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] h-[40px] text-center border border-borderColor px-3  rounded-lg   font-AeonikProRegular "
-                  placeholder="Мин"
-                />
-              </div>
-              <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-[60px] h-[40px] text-center border border-borderColor px-3  rounded-lg  font-AeonikProRegular "
-                  placeholder="Макс"
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-        <div className="w-full flex flex-row px-3 pt-5 gap-x-[11px] md:gap-x-[20px] mb-[15px]">
-          <div className="w-fit flex items-center gap-x-[25px]">
-            <div className="w-fit hidden md:flex flex-col items-start">
-              <div className="flex items-center justify-center  mb-2 ll:mb-[10px]">
-                <label
-                  htmlFor=""
-                  className="flex items-center text-[14px] ll:text-base text-mobileTextColor  ll:font-AeonikProMedium font-AeonikProRegular">
-
-                  Возраст
-                </label>
-                {/* <span className="ml-[5px]">
-                  <StarLabel />
-                </span> */}
-              </div>
-              <div className="w-fit flex items-center">
-                <input
-                  type="text"
-                  className="inputStyle w-[58px] h-[40px] text-center fon border border-borderColor rounded-lg px-[12px]  outline-none "
-                  placeholder=""
-                />
-              </div>
-            </div>
-            <div className="w-full md:w-[55%]">
-              <div className="flex items-center mb-2 ll:mb-[10px] ">
-                <label
-                  htmlFor=""
-                  className="flex items-center text-[14px] ll:text-base text-mobileTextColor  ll:font-AeonikProMedium font-AeonikProRegular">
-
-                  Цена
-                </label>
-                <span className="ml-[5px]">
-                  <StarLabel />
-                </span>
-              </div>
-              <div className="w-full h-[40px] flex items-center border border-borderColor px-3 py-[6px] rounded-lg text-xs ">
-                <input
-                  type="text"
-                  placeholder="0"
-                  className="inputStyle w-[70%] font-AeonikProMedium outline-none "
-                />
-                <span className="text-textLightColor ml-[10px] text-xs md:text-base font-AeonikProRegular">
-                  сум
-                </span>
-              </div>
-            </div>
-          </div>
-          <div className="w-fit flex flex-col items-start">
-            <div className="flex items-center justify-center mb-2 ll:mb-[10px] ">
-              <label
-                htmlFor=""
-                className="flex items-center text-[14px] ll:text-base text-mobileTextColor  ll:font-AeonikProMedium font-AeonikProRegular">
-
-                Скидка
-              </label>
-              <span className="font-AeonikProMedium text-[10px] md:text-[13px] text-textLightColor ml-[5px]">
-                (необязательно)
-              </span>
-            </div>
-            <div className="w-full flex items-center justify-center">
-              <div className="w-full flex items-center gap-x-1">
-                <div className="w-[40%] md:w-[72px] flex items-start">
-                  <div className="w-full h-10 flex items-center justify-center border border-borderColor rounded-lg px-[10px] md:px-3 py-[8px]">
-                    <input
-                      type="text"
-                      placeholder="0"
-                      className="inputStyle w-[70%] font-AeonikProMedium text-start outline-none "
-                    />
-                    <span className="text-textLightColor ml-2">%</span>
-                  </div>
-                </div>
-                <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-                <div className="w-[60%] md:w-[75%] flex items-center">
-                  <div className="w-full h-[40px] flex items-center justify-between border border-borderColor px-3 py-[6px] rounded-lg text-xs">
-                    <input
-                      type="text"
-                      placeholder="0"
-                      className="inputStyle w-[75%] font-AeonikProMedium outline-none "
-                    />
-                    <span className="text-textLightColor ml-[10px] text-xs md:text-base font-AeonikProRegular">
-                      сум
-                    </span>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <button className="w-full flex items-end justify-end text-lg text-textBlueColor font-AeonikProMedium pr-1">
-          готово
-        </button>
-      </div>
-    </div>
-  );
-  // Accessories bor
-  const contentAccessories = (
-    <div className="w-[595px] h-fit">
-      <div
-        className={`w-full h-fit flex flex-col cursor-pointer font-AeonikProMedium text-sm leading-4 text-center hover:bg-bgColor`}
-      >
-        <div className="w-full flex gap-x-10 px-3 pt-5">
-          <div className="w-fit flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Размер{" "}
-              <span className="text-sm text-textLightColor ml-[6px]">(см)</span>
-            </p>
-            <div className="w-[83px] flex items-center justify-between gap-x-1">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-full text-start h-[38px] border border-borderColor px-3 rounded-lg  font-AeonikProRegular "
-                />
-              </div>
-            </div>
-          </div>
-
-          <div className="w-[80%] flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-              Обхват Размер
-            </p>
-            <AllSizeListForWear />
-          </div>
-        </div>
-        <div className="w-full flex gap-x-10 px-3 pt-5">
-          <div className="w-[20%] flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Длина
-              <span className="text-sm text-textLightColor ml-[6px]">(см)</span>
-            </p>
-            <div className="flex items-center justify-between">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-full h-[40px] text-start border border-borderColor px-3 rounded-lg   font-AeonikProRegular "
-                />
-              </div>
-            </div>
-          </div>
-          <div className="w-[20%] flex flex-col">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Ширина
-            </p>
-            <div className="flex items-center justify-between gap-x-1">
-              <div className="flex flex-col">
-                <input
-                  type="text"
-                  className="inputStyle w-full h-[40px] text-start border border-borderColor px-3  rounded-lg  font-AeonikProRegular "
-                />
-              </div>
-            </div>
-          </div>
-          <div className="w-[60%] flex flex-col ml-auto">
-            <p className="flex items-center text-[14px] ll:text-base text-mobileTextColor mb-2 ll:mb-[10px] ll:font-AeonikProMedium font-AeonikProRegular">
-
-              Количество
-            </p>
-            <div className="flex items-start justify-between ">
-              <input
-                type="text"
-                className="inputStyle w-[60px] h-[40px] text-center border border-borderColor px-5  rounded-lg  font-AeonikProRegular "
-              />
-            </div>
-          </div>
-        </div>
-        <div className="w-full flex flex-row px-3 pt-5 gap-x-[11px] md:gap-x-[20px] mb-[15px]">
-          <div className="w-[45%] flex items-center gap-x-[25px]">
-            <div className="w-fit hidden md:flex flex-col items-start">
-              <div className="flex items-center justify-center mb-2 ll:mb-[10px] ">
-                <label
-                  htmlFor=""
-                  className="flex items-center text-[14px] ll:text-base text-mobileTextColor ll:font-AeonikProMedium font-AeonikProRegular">
-
-                  Возраст
-                </label>
-                {/* <span className="ml-[5px]">
-                  <StarLabel />
-                </span> */}
-              </div>
-              <div className="w-fit flex items-center">
-                <input
-                  type="text"
-                  className="inputStyle w-[58px] h-[42px] text-center fon border border-borderColor rounded-lg px-[12px] outline-none "
-                  placeholder=""
-                />
-              </div>
-            </div>
-            <div className="w-full md:w-[55%]">
-              <div className="flex items-center mb-2 ll:mb-[10px] ">
-                <label
-                  htmlFor=""
-                  className="flex items-center text-[14px] ll:text-base text-mobileTextColor  ll:font-AeonikProMedium font-AeonikProRegular">
-
-                  Цена
-                </label>
-                <span className="ml-[5px]">
-                  <StarLabel />
-                </span>
-              </div>
-              <div className="w-full h-[40px] flex items-center border border-borderColor px-3 py-[6px] rounded-lg text-xs">
-                <input
-                  type="text"
-                  placeholder="0"
-                  className="inputStyle w-[70%] font-AeonikProMedium outline-none "
-                />
-                <span className="text-textLightColor ml-[10px] text-xs md:text-base font-AeonikProRegular">
-                  сум
-                </span>
-              </div>
-            </div>
-          </div>
-          <div className="w-[40%] flex flex-col items-start">
-            <div className="flex items-center justify-center mb-2 ll:mb-[10px] ">
-              <label
-                htmlFor=""
-                className="flex items-center text-[14px] ll:text-base text-mobileTextColor ll:font-AeonikProMedium font-AeonikProRegular">
-
-                Скидка
-              </label>
-              <span className="font-AeonikProMedium text-[10px] md:text-[13px] text-textLightColor ml-[5px]">
-                (необязательно)
-              </span>
-            </div>
-            <div className="w-full flex items-center justify-center">
-              <div className="w-full flex items-center gap-x-1">
-                <div className="w-[40%] md:w-[72px] flex items-start">
-                  <div className="w-full h-10 flex items-center justify-center border border-borderColor rounded-lg px-[10px] md:px-3 py-[8px]">
-                    <input
-                      type="text"
-                      placeholder="0"
-                      className="inputStyle w-[70%] font-AeonikProMedium text-start outline-none "
-                    />
-                    <span className="text-textLightColor ml-2">%</span>
-                  </div>
-                </div>
-                <span className="w-[15px] h-[2px] bg-borderColor  mx-[4px]"></span>
-                <div className="w-[60%] md:w-[75%] flex items-center">
-                  <div className="w-full h-[40px] flex items-center justify-between border border-borderColor px-3 py-[6px] rounded-lg text-xs">
-                    <input
-                      type="text"
-                      placeholder="0"
-                      className="inputStyle w-[75%] font-AeonikProMedium outline-none "
-                    />
-                    <span className="text-textLightColor ml-[10px] text-xs md:text-base font-AeonikProRegular">
-                      сум
-                    </span>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <button className="w-full flex items-end justify-end text-lg text-textBlueColor font-AeonikProMedium pr-1">
-          готово
-        </button>
-      </div>
-    </div>
-  );
 
   // const changeColor = [
   //   { id: 1, data: 1, icons: InputCheck, action: false, colors: "bg-black" },
@@ -1291,7 +371,7 @@ const AddingProduct = () => {
                       onClick={() =>
                         HandleIconsColor(data?.IconsColor, data?.id)
                       }
-                      className={`rounded-[12px] flex items-center justify-center mr-2 w-[65px] h-[40px] bg-[#aaaaaa] bg-[${data.hex}] cursor-pointer ${data?.id == 2 ? "border border-setTexOpacity flex items-center justify-center" : "" }
+                      className={`rounded-[12px] flex items-center justify-center mr-2 w-[65px] h-[40px] bg-[#aaaaaa] bg-[${data.hex}] cursor-pointer ${data?.id == 2 ? "border border-setTexOpacity flex items-center justify-center" : ""}
                      `}
                     >
                       {data?.action && data?.id === 2 ? (
@@ -1460,7 +540,7 @@ const AddingProduct = () => {
             <div className="w-1/4 flex flex-col">
               <div className="w-full h-[95px] flex items-center justify-center rounded-lg border border-dashed border-borderColor mb-2">
                 <Link to="#">
-                 <DownloadIcon />
+                  <DownloadIcon />
                 </Link>
               </div>
               <div className="w-full h-[95px] flex flex-col items-center justify-center rounded-lg border border-dashed border-borderColor mb-2">
@@ -1976,61 +1056,11 @@ const AddingProduct = () => {
               {/* 5 Категория одежды*/}
               {state.openDropModalButton ? (
                 <div className="w-full hidden md:flex items-center flex-wrap gap-3 ">
-                  <Popover
-                    // open={state?.openhat}
-                    // onOpenChange={handleOpenChangeHat}
-                    className="px-[15px] h-[38px] border-textBlueColor text-textBlueColor border-[1.5px] font-AeonikProMedium flex items-center text-sm justify-center cursor-pointer active:scale-95 rounded-lg focus:bg-textBlueColor hover:bg-textBlueColor hover:text-white transition duration-300"
-                    trigger="click"
-                    options={["Hide"]}
-                    placement="bottomLeft"
-                    content={contentHat}
-                  >
-                    Головные уборы
-                  </Popover>
-                  <Popover
-                    // open={state?.openOutwear}
-                    // onOpenChange={handleOpenChangeOutwear}
-                    className="group px-[15px] h-[38px] border-textBlueColor text-textBlueColor border-[1.5px] font-AeonikProMedium flex items-center text-sm justify-center cursor-pointer active:scale-95 rounded-lg hover:bg-textBlueColor hover:text-white transition duration-300"
-                    trigger="click"
-                    options={["Hide"]}
-                    placement="bottomLeft"
-                    content={contentOutwear}
-                  >
-                    Верхняя одежда
-                  </Popover>
-                  <Popover
-                    // open={state?.openOutwaer}
-                    // onOpenChange={handleOpenChangeOutwear}
-                    className="group px-[15px] h-[38px] rounded-lg border-textBlueColor text-textBlueColor border-[1.5px] font-AeonikProMedium flex items-center text-sm justify-center cursor-pointer active:scale-95 focus:bg-textBlueColor focus:text-white hover:bg-textBlueColor hover:text-white transition duration-300"
-                    trigger="click"
-                    options={["Hide"]}
-                    placement="bottom"
-                    content={contentUnderWear}
-                  >
-                    Нижняя одежда
-                  </Popover>
-                  <Popover
-                    // open={state?.openwear}
-                    // onOpenChange={handleOpenChangeWear}
-                    className="px-[15px] h-[38px] border-textBlueColor text-textBlueColor border-[1.5px] font-AeonikProMedium flex items-center justify-center text-sm cursor-pointer active:scale-95 rounded-lg focus:bg-textBlueColor focus:text-white hover:bg-textBlueColor hover:text-white transition duration-300"
-                    trigger="click"
-                    options={["Hide"]}
-                    placement="bottom"
-                    content={contentShoes}
-                  >
-                    Обувь
-                  </Popover>
-                  <Popover
-                    // open={state?.openwear}
-                    // onOpenChange={handleOpenChangeWear}
-                    className="group px-[15px] h-[38px] border-textBlueColor text-textBlueColor border-[1.5px] font-AeonikProMedium flex items-center justify-center text-sm cursor-pointer active:scale-95  rounded-lg focus:bg-textBlueColor focus:text-white hover:bg-textBlueColor hover:text-white transition duration-300"
-                    trigger="click"
-                    options={["Hide"]}
-                    placement="bottomRight"
-                    content={contentAccessories}
-                  >
-                    Аксессуары
-                  </Popover>
+                  <HeadWearAdd />
+                  <OutWearAdd />
+                  <UnderWearAdd />
+                  <ShoesAdd />
+                  <AccessoriesAdd />
                 </div>
               ) : (
                 ""
